+++ src/features/app-updates/components/WebUIUpdateChecker.tsx
@@ -14,13 +14,12 @@ import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
 import DialogActions from '@mui/material/DialogActions';
 import Button from '@mui/material/Button';
-import { d } from 'koration';
 import { defaultPromiseErrorHandler } from '@/lib/DefaultPromiseErrorHandler.ts';
 import { UpdateState, WebUiChannel, WebUiUpdateStatus } from '@/lib/graphql/generated/graphql.ts';
 import { useLocalStorage, useSessionStorage } from '@/base/hooks/useStorage.tsx';
 import { requestManager } from '@/lib/requests/RequestManager.ts';
 import { makeToast } from '@/base/utils/Toast.ts';
-import { ABOUT_WEBUI, WEBUI_UPDATE_CHECK } from '@/lib/graphql/fragments/InfoFragments.ts';
+import { ABOUT_WEBUI, WEBUI_UPDATE_CHECK } from '@/lib/graphql/server/InfoFragments.ts';
 import { VersionUpdateInfoDialog } from '@/features/app-updates/components/VersionUpdateInfoDialog.tsx';
 import { useUpdateChecker } from '@/features/app-updates/hooks/useUpdateChecker.tsx';
 import { useMetadataServerSettings } from '@/features/settings/services/ServerSettingsMetadata.ts';
@@ -29,8 +28,6 @@ import { AppStorage } from '@/lib/storage/AppStorage.ts';
 import { BrowserUtil } from '@/lib/BrowserUtil.ts';
 
 const disabledUpdateCheck = () => Promise.resolve();
-
-const FORCED_REFRESH_THRESHOLD = d(30).seconds.inWholeMilliseconds;
 
 const INITIAL_LOAD_TIMESTAMP_KEY = 'webUIInitialLoadTimestamp';
 
@@ -82,9 +79,7 @@ export const WebUIUpdateChecker = () => {
     const newVersion = aboutWebUI?.tag;
     const isSameAsCurrent = !newVersion || !webUIVersion || webUIVersion === newVersion;
 
-    // Calculate if forced refresh threshold has been met
-    const timeSinceLoad = Date.now() - (initialLoadTimestamp ?? Date.now());
-    const shouldForceRefresh = timeSinceLoad >= FORCED_REFRESH_THRESHOLD;
+    const shouldForceRefresh = initialLoadTimestamp < Number(aboutWebUI?.updateTimestamp);
 
     const saveInitialVersion = !webUIVersion && !!newVersion;
     if (saveInitialVersion) {