+++ src/features/settings/Settings.constants.ts
@@ -9,7 +9,13 @@
 import { d } from 'koration';
 import { DEFAULT_DEVICE } from '@/features/device/services/Device.ts';
 import { DEFAULT_SORT_SETTINGS } from '@/features/migration/Migration.constants.ts';
-import { GlobalUpdateSkipEntriesSettings, MetadataServerSettings } from '@/features/settings/Settings.types.ts';
+import {
+    GlobalUpdateSkipEntriesSettings,
+    ImageProcessingTargetMode,
+    ImageProcessingType,
+    MetadataServerSettings,
+    ServerSettings,
+} from '@/features/settings/Settings.types.ts';
 import { GridLayout, TranslationKey } from '@/base/Base.types.ts';
 import { getDefaultLanguages } from '@/base/utils/Languages.ts';
 import { SelectSettingValue, SelectSettingValueDisplayInfo } from '@/base/components/settings/SelectSetting.tsx';
@@ -256,3 +262,56 @@ export const KOREADER_SYNC_CHECKSUM_METHOD_SELECT_VALUES: SelectSettingValue<Kor
         method,
         KOREADER_SYNC_CHECKSUM_METHOD_TO_TRANSLATION_KEYS[method],
     ]);
+
+export const IMAGE_PROCESSING_COMPRESSION = {
+    min: 0,
+    max: 1,
+    step: 0.01,
+};
+export const IMAGE_PROCESSING_CALL_TIMEOUT = {
+    min: d(10).seconds.inWholeSeconds,
+    max: d(10).minutes.inWholeSeconds,
+    step: d(10).seconds.inWholeSeconds,
+};
+export const IMAGE_PROCESSING_CONNECT_TIMEOUT = {
+    min: d(10).seconds.inWholeSeconds,
+    max: d(10).minutes.inWholeSeconds,
+    step: d(10).seconds.inWholeSeconds,
+};
+
+const IMAGE_PROCESSING_TARGET_MODES = Object.values(ImageProcessingTargetMode);
+const IMAGE_PROCESSING_TARGET_MODES_TO_TRANSLATION_KEY: {
+    [flavor in ImageProcessingTargetMode]: SelectSettingValueDisplayInfo;
+} = {
+    [ImageProcessingTargetMode.DISABLED]: {
+        text: 'global.label.disabled',
+    },
+    [ImageProcessingTargetMode.IMAGE]: {
+        text: 'download.settings.conversion.target_modes.image.title',
+        description: 'download.settings.conversion.target_modes.image.description',
+    },
+    [ImageProcessingTargetMode.URL]: {
+        text: 'download.settings.conversion.target_modes.url.title',
+        description: 'download.settings.conversion.target_modes.image.description',
+    },
+};
+export const IMAGE_PROCESSING_TARGET_MODES_SELECT_VALUES: SelectSettingValue<ImageProcessingTargetMode>[] =
+    IMAGE_PROCESSING_TARGET_MODES.map((mode) => [mode, IMAGE_PROCESSING_TARGET_MODES_TO_TRANSLATION_KEY[mode]]);
+
+export const IMAGE_PROCESSING_INPUT_WIDTH = 250;
+export const DEFAULT_MIME_TYPE = 'default';
+export const MIME_TYPE_PREFIX = 'image/';
+export const TARGET_DISABLED = 'none';
+
+export const IMAGE_PROCESSING_TYPE_TO_TRANSLATION: Record<ImageProcessingType, TranslationKey> = {
+    [ImageProcessingType.DOWNLOAD]: 'download.settings.conversion.title',
+    [ImageProcessingType.SERVE]: 'settings.images.processing.serve.title',
+};
+
+export const IMAGE_PROCESSING_TYPE_TO_SETTING: Record<
+    ImageProcessingType,
+    Extract<keyof ServerSettings, 'downloadConversions' | 'serveConversions'>
+> = {
+    [ImageProcessingType.DOWNLOAD]: 'downloadConversions',
+    [ImageProcessingType.SERVE]: 'serveConversions',
+};